Boxfresh Grocery Website
The Online Grocery Store Boxfresh is developed to override the problems faced by the Grocery Stores in practicing the manual system, reduce the hardships faced by the existing system and to carry out operations in a smooth and effective manner. It also generates an error message on entering invalid data. No formal knowledge is required to use this system. It is user-friendly. Users can concentrate on their other works rather than record keeping. Thus, it helps organization in better utilization of resources.